+++ R/decode.R
@@ -49,8 +49,9 @@ setAs(from = "corpus", to = "Annotation", def = function(from){
 
 #' Decode corpus or subcorpus.
 #' 
-#' Decode \code{corpus} or \code{subcorpus} and return class specified by argument \code{to}.
-#' 
+#' Decode \code{corpus} or \code{subcorpus} and return class specified by
+#' argument \code{to}.
+#'
 #' The primary purpose of the method is type conversion. By obtaining the corpus
 #' or subcorpus in the format specified by the argument \code{to}, the data can
 #' be processed with tools that do not rely on the Corpus Workbench (CWB).
@@ -59,21 +60,27 @@ setAs(from = "corpus", to = "Annotation", def = function(from){
 #' defined in the package \code{NLP}. Another purpose of decoding the corpus can
 #' be to rework it, and to re-import it into the CWB (e.g. using the
 #' \code{cwbtools}-package).
-#' 
+#'
 #' An earlier version of the method included an option to decode a single
-#' s-attribute, which is not supported any more. See the \code{s_attribute_decode}
-#' function of the package RcppCWB.
-#' 
-#' @return The return value will correspond to the class specified by argument \code{to}. 
+#' s-attribute, which is not supported any more. See the
+#' \code{s_attribute_decode} function of the package RcppCWB.
+#'
+#' @return The return value will correspond to the class specified by argument
+#'   \code{to}.
 #' 
 #' @param .Object The \code{corpus} or \code{subcorpus} to decode.
 #' @param to The class of the returned object, stated as a length-one
 #'   \code{character} vector.
-#' @param s_attributes The structural attributes to decode. If \code{NULL} (default), all
-#'   structural attributes will be decoded.
-#' @param p_attributes The positional attributes to decode. If \code{NULL} (default), all
-#'   positional attributes will be decoded.
+#' @param s_attributes The structural attributes to decode. If \code{NULL}
+#'   (default), all structural attributes will be decoded.
+#' @param p_attributes The positional attributes to decode. If \code{NULL}
+#'   (default), all positional attributes will be decoded.
+#' @param decode A \code{logical} value, whether to decode token ids and struc
+#'   ids to character strings. If \code{FALSE}, the values of columns for p- and
+#'   s-attributes will be \code{integer} vectors. If \code{TRUE} (default), the
+#'   respective columns are \code{character} vectors.
 #' @param ... Further arguments.
+#' @param verbose A \code{logical} value, whether to output progess messages.
 #' @exportMethod decode
 #' @importFrom RcppCWB get_region_matrix
 #' @rdname decode
@@ -104,11 +111,9 @@ setAs(from = "corpus", to = "Annotation", def = function(from){
 #' 
 #' # Previous versions of polmineR offered an option to decode a single
 #' # s-attribute. This is how you could proceed to get a table with metadata.
+#' dt <- decode(P, s_attribute = "id", decode = FALSE)
 #' dt[, "word" := NULL]
-#' dt[,
-#'   {list(cpos_left = min(.SD[["cpos"]]), cpos_right = max(.SD[["cpos"]]), id = unique(.SD[["id"]]))},
-#'   by = "struc"
-#'   ]
+#' dt[,{list(cpos_left = min(.SD[["cpos"]]), cpos_right = max(.SD[["cpos"]]))}, by = "id"]
 #' 
 #' # Decode subcorpus as Annotation object
 #' \dontrun{
@@ -127,35 +132,48 @@ setAs(from = "corpus", to = "Annotation", def = function(from){
 #' }
 #' }
 #' @rdname decode
-setMethod("decode", "corpus", function(.Object, to = c("data.table", "Annotation"), p_attributes = NULL, s_attributes = NULL){
+setMethod("decode", "corpus", function(.Object, to = c("data.table", "Annotation"), p_attributes = NULL, s_attributes = NULL, decode = TRUE, verbose = TRUE){
   if (to == "data.table"){
 
     if (is.null(p_attributes)) p_attributes <- p_attributes(.Object)
     if (!all(p_attributes %in% p_attributes(.Object))) stop("Not all p_attributes provided are available.")
     
-    if (is.null(s_attributes)) s_attributes <- s_attributes(.Object)
-    if (!all(s_attributes %in% s_attributes(.Object))) stop("Not all s_attributes provided are available.")
 
     p_attribute_list <- lapply(
       setNames(p_attributes, p_attributes),
       function(p_attr){
-        message("decoding p-attribute:", p_attr)
-        get_token_stream(.Object, p_attribute = p_attr)
+        if (verbose) message("decoding p-attribute:", p_attr)
+        get_token_stream(.Object, p_attribute = p_attr, decode = decode)
       }
     )
 
     max_cpos <- size(.Object) - 1L
-    s_attribute_list <- lapply(
-      setNames(s_attributes, s_attributes),
-      function(s_attr){
-        message("decoding s-attribute:", s_attr)
-        corpus_id <- get_corpus(.Object)
-        struc <- cl_cpos2struc(corpus = corpus_id, s_attribute = s_attr, cpos = 0L:max_cpos, registry = registry())
-        str <- cl_struc2str(corpus = corpus_id, s_attribute = s_attr, struc = struc, registry = registry())
-        Encoding(str) <- encoding(.Object)
-        as.nativeEnc(str, from = encoding(.Object))
-      }
-    )
+    
+    if (is.null(s_attributes)) s_attributes <- s_attributes(.Object)
+    if (length(s_attributes) > 0L){
+      if (!all(s_attributes %in% s_attributes(.Object))) stop("Not all s_attributes provided are available.")
+    }
+    
+    if (length(s_attributes) > 0L){
+      s_attribute_list <- lapply(
+        setNames(s_attributes, s_attributes),
+        function(s_attr){
+          if (verbose) message("decoding s-attribute:", s_attr)
+          corpus_id <- get_corpus(.Object)
+          struc <- cl_cpos2struc(corpus = corpus_id, s_attribute = s_attr, cpos = 0L:max_cpos, registry = registry())
+          if (decode){
+            str <- cl_struc2str(corpus = corpus_id, s_attribute = s_attr, struc = struc, registry = registry())
+            Encoding(str) <- encoding(.Object)
+            return(as.nativeEnc(str, from = encoding(.Object)))
+          } else {
+            return(struc)
+          }
+        }
+      )
+    } else {
+      s_attribute_list <- list()
+    }
+    
 
     message("assembling data.table")
     combined_list <- c(
@@ -173,14 +191,15 @@ setMethod("decode", "corpus", function(.Object, to = c("data.table", "Annotation
   y
 })
 
+
 #' @exportMethod decode
 #' @rdname decode
-setMethod("decode", "character", function(.Object, to = c("data.table", "Annotation"), s_attributes = NULL, p_attributes = NULL, ...){
+setMethod("decode", "character", function(.Object, to = c("data.table", "Annotation"), s_attributes = NULL, p_attributes = NULL, decode = TRUE, verbose = TRUE, ...){
   if (any(c("sAttribute", "s_attribute") %in% names(list(...)))){
     stop("Decoding an s_attribute is not supported any longer in the decode()-method of ",
          "the polmineR package. See s_attribute_decode in the RcppCWB package as a substitute.")
   }
-  decode(corpus(.Object), to = to, s_attributes = s_attributes, p_attributes = p_attributes)
+  decode(corpus(.Object), to = to, s_attributes = s_attributes, p_attributes = p_attributes, decode = decode, verbose = verbose)
 })
 
 
@@ -189,60 +208,54 @@ setMethod("decode", "character", function(.Object, to = c("data.table", "Annotat
 
 #' @exportMethod decode
 #' @rdname decode
-setMethod("decode", "slice", function(.Object, to = "data.table", s_attributes = NULL, p_attributes = NULL){
+setMethod("decode", "slice", function(.Object, to = "data.table", s_attributes = NULL, p_attributes = NULL, decode = TRUE, verbose = TRUE){
   if (to == "data.table"){
     
     if (is.null(p_attributes)) p_attributes <- p_attributes(.Object)
     if (!all(p_attributes %in% p_attributes(.Object))) stop("Not all p_attributes provided are available.")
     
-    if (is.null(s_attributes)) s_attributes <- s_attributes(.Object)
-    if (!all(s_attributes %in% s_attributes(.Object))) stop("Not all s_attributes provided are available.")
+    y <- data.table(cpos = unlist(apply(.Object@cpos, 1, function(row) row[1]:row[2])))
 
-    ts <- lapply(
-      setNames(p_attributes, p_attributes),
-      function(p_attr){
-        message("... decoding p_attribute ", p_attr)
-        get_token_stream(.Object, p_attribute = p_attr)
-      }
-    )
-    p_attr_dt <- as.data.table(ts)
-    p_attr_dt[, "cpos" := unlist(apply(.Object@cpos, 1, function(row) row[1]:row[2]))]
-    
-    strucs <- RcppCWB::cl_cpos2struc(corpus = .Object@corpus, s_attribute = s_attributes[1], cpos = .Object@cpos[,1])
-    
-    dts <- lapply(
-      strucs,
-      function(struc){
-        region <- RcppCWB::cl_struc2cpos(corpus = .Object@corpus, s_attribute = s_attributes[1], struc = struc)
-        data.table(struc = struc, cpos_left = region[1], cpos_right = region[2])
-      }
-    )
-    regions <- rbindlist(dts)
-    
-    s_attr_values <- lapply(
-      setNames(s_attributes, s_attributes),
-      function(s_attr){
-        message("... decoding s_attribute ", s_attr)
-        str <- RcppCWB::cl_struc2str(corpus = .Object@corpus, s_attribute = s_attr, struc = strucs)
-        Encoding(str) <- encoding(.Object)
-        as.nativeEnc(str, from = encoding(.Object))
-      }
-    )
-    dt <- as.data.table(s_attr_values)
-    
-    s_attr_dt <- cbind(regions, dt)
-
-    unfold <- function(.SD){
-      dt <- data.table(cpos = .SD[["cpos_left"]]:.SD[["cpos_right"]])
-      for (s_attr in s_attributes) dt[[s_attr]] <- rep(.SD[[s_attr]], times = nrow(dt))
-      dt
+    for (p_attr in p_attributes){
+      if (verbose) message("... decoding p_attribute ", p_attr)
+      y[, (p_attr) := get_token_stream(.Object, p_attribute = p_attr, decode = decode)]
     }
-    s_attr_dt_ext <- s_attr_dt[, unfold(.SD), by = "struc"]
     
-    setkeyv(p_attr_dt, cols = "cpos")
-    setkeyv(s_attr_dt_ext, cols = "cpos")
-    y <- p_attr_dt[s_attr_dt_ext]
-    
+    if (is.null(s_attributes)) s_attributes <- s_attributes(.Object)
+    if (length(s_attributes) > 0L){
+      if (!all(s_attributes %in% s_attributes(.Object))) stop("Not all s_attributes provided are available.")
+      strucs <- RcppCWB::cl_cpos2struc(corpus = .Object@corpus, s_attribute = s_attributes[1], cpos = .Object@cpos[,1])
+      
+      s_attr_dt <- data.table(
+        RcppCWB::get_region_matrix(corpus = .Object@corpus, s_attribute = s_attributes[1], strucs = strucs, registry = registry())
+      )
+      setnames(s_attr_dt, old = c("V1", "V2"), new = c("cpos_left", "cpos_right"))
+      s_attr_dt[, "struc" := strucs]
+      
+      for (s_attr in s_attributes){
+        if (decode){
+          if (verbose) message("... decoding s_attribute ", s_attr)
+          str <- RcppCWB::cl_struc2str(corpus = .Object@corpus, s_attribute = s_attr, struc = strucs)
+          Encoding(str) <- encoding(.Object)
+          s_attr_dt[, (s_attr) := as.nativeEnc(str, from = encoding(.Object))]
+        } else {
+          s_attr_dt[, (s_attr) := strucs]
+        }
+      }
+      
+      unfold <- function(.SD){
+        dt <- data.table(cpos = .SD[["cpos_left"]]:.SD[["cpos_right"]])
+        value <- .SD[[s_attr]]
+        for (s_attr in s_attributes) dt[, (s_attr) := value]
+        dt
+      }
+      s_attr_dt_ext <- s_attr_dt[, unfold(.SD), by = "struc"]
+      setkeyv(y, cols = "cpos")
+      setkeyv(s_attr_dt_ext, cols = "cpos")
+      y <- y[s_attr_dt_ext]
+      y[, "struc" := NULL]
+      setcolorder(y, neworder = c("cpos", p_attributes, s_attributes))
+    }
   } else if (to == "Annotation"){
     y <- as(.Object, "Annotation")
   }
@@ -250,14 +263,14 @@ setMethod("decode", "slice", function(.Object, to = "data.table", s_attributes =
 })
 
 #' @rdname decode
-setMethod("decode", "partition", function(.Object, to = "data.table", s_attributes = NULL, p_attributes = NULL){
+setMethod("decode", "partition", function(.Object, to = "data.table", s_attributes = NULL, p_attributes = NULL, decode = TRUE, verbose = TRUE){
   callNextMethod()
 })
 
 
 #' @exportMethod decode
 #' @rdname decode
-setMethod("decode", "subcorpus", function(.Object, to = "data.table", s_attributes = NULL, p_attributes = NULL){
+setMethod("decode", "subcorpus", function(.Object, to = "data.table", s_attributes = NULL, p_attributes = NULL, decode = TRUE, verbose = TRUE){
   callNextMethod()
 })