+++ R/utils.R
@@ -320,3 +320,4 @@ round.data.table <- function(x, digits = 2L){
 #' @importFrom magrittr %>%
 #' @export
 magrittr::`%>%`
+